small cleanup and extra check for type of Sections tag

+++ src/main/java/net/querz/mcaselector/filter/LongFilter.java
@@ -39,12 +39,12 @@ public abstract class LongFilter extends NumberFilter<Long> {
 
 	@Override
 	boolean isEqual(Long a, Long b) {
-		return a.intValue() == b.intValue();
+		return a.longValue() == b.longValue();
 	}
 
 	@Override
 	boolean isNotEqual(Long a, Long b) {
-		return a.intValue() != b.intValue();
+		return a.longValue() != b.longValue();
 	}
 
 	@Override
